+++ src/ram.cpp
@@ -1,6 +1,6 @@
 #include "ram.h"
 
-void RAM::Init(const int alloc) {
+void RAM::Init(const long long unsigned int alloc) {
     this->memory = new byte[alloc];
 }
